refactor: move `ColorMode` to `iostreams` to avoid cyclical dependencies and resolve the color output based on the custom type
